make tvshows available
